requirements.txt
transformers==4.38.2
peft==0.10.0
torch==2.2.1
python-igraph==0.9.6
pyautogen>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
biopython>=1.79
requests>=2.25.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
loguru>=0.5.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.5b0
flake8>=3.9.0
